Conversion display successful
Removed old research print and RSS info requests to a getInputData function.

Enabled a loop to search more than once for a user.

Nailed down index-based RSS feed tag data for the description and itemList in findConversion(..) to accurately display the conversion countries, codes, and values.

Added an import time - will be used to automate a time-based refresh of the user's chosen currency.